refactor: move "Challenge*" from "halo2_common" to "halo2_backend"
